# Bunch of speed and memory optimizations