--- a/lib/i18n/locale/tag/parents.rb
+++ b/lib/i18n/locale/tag/parents.rb
@@ -3,18 +3,20 @@ module I18n
     module Tag
       module Parents
         def parent
-          @parent ||= begin
-            segs = to_a.compact
-            segs.length > 1 ? self.class.tag(*segs[0..(segs.length-2)].join('-')) : nil
-          end
+          @parent ||=
+            begin
+              segs = to_a
+              segs.compact!
+              segs.length > 1 ? self.class.tag(*segs[0..(segs.length - 2)].join('-')) : nil
+            end
         end
 
         def self_and_parents
-          @self_and_parents ||= [self] + parents
+          @self_and_parents ||= [self].concat parents
         end
 
         def parents
-          @parents ||= ([parent] + (parent ? parent.parents : [])).compact
+          @parents ||= parent ? [parent].concat(parent.parents) : []
         end
       end
     end