`I18n::Locale::Fallbacks.new[locale]` is faster around 30% and uses around 40% less memory

```ruby
I18n::Locale::Fallbacks.new[:de]

Before:
Calculating -------------------------------------
            original     1.696k memsize (     0.000  retained)
                        36.000  objects (     0.000  retained)
                         2.000  strings (     0.000  retained)

Warming up --------------------------------------
            original     9.584k i/100ms
Calculating -------------------------------------
            original    100.989k (± 2.9%) i/s -    507.952k in   5.034224s


After:
Calculating -------------------------------------
            original     1.056k memsize (     0.000  retained)
                        20.000  objects (     0.000  retained)
                         2.000  strings (     0.000  retained)

Warming up --------------------------------------
            original    13.143k i/100ms
Calculating -------------------------------------
            original    135.522k (± 3.4%) i/s -    683.436k in   5.049384s
```

```ruby
I18n::Locale::Fallbacks.new[:'en-US']

Before:
Calculating -------------------------------------
            original     2.496k memsize (     0.000  retained)
                        55.000  objects (     0.000  retained)
                         4.000  strings (     0.000  retained)

ips benchmark:
Warming up --------------------------------------
            original     6.597k i/100ms
Calculating -------------------------------------
            original     68.147k (± 4.8%) i/s -    343.044k in   5.046789s


After:
Calculating -------------------------------------
            original     1.696k memsize (     0.000  retained)
                        35.000  objects (     0.000  retained)
                         4.000  strings (     0.000  retained)

ips benchmark:
Warming up --------------------------------------
            original     8.169k i/100ms
Calculating -------------------------------------
            original     87.410k (± 2.7%) i/s -    441.126k in   5.050400s
```